+++ rlapy/drivers/least_squares.py
@@ -3,9 +3,6 @@ Routines for (approximately) solving over-determined least squares problems
 
     min{ || A x - b ||_2 + reg || x ||_2 : x in R^n }.
 """
-#TODO: implement LSRN's solver for strongly under-determined problems.
-#TODO: implement Blendenpik's solver for strongly under-determined problems
-#  (that implementation is only in the Matlab code, not in the paper).
 import warnings
 import scipy.linalg as la
 import numpy as np
@@ -199,63 +196,64 @@ class SPO3(OverLstsqSolver):
             assert tol >= 0
             assert tol < np.inf
         rng = np.random.default_rng(rng)
+        
+        if logging:
+            quick_time = time.time()
+        else:
+            quick_time = lambda: 0.0
 
         # Sketch the data matrix
-        tic = time.time() if logging else 0
+        tic = quick_time()
         S = self.sketch_op_gen(d, n_rows, rng)
         A_ske = S @ A
-        toc = time.time() if logging else 0
-        time_sketch = toc - tic
-        self.log['time_sketch'] = time_sketch
+        self.log['time_sketch'] = quick_time() - tic
 
         # Factor the sketch
-        tic = time.time() if logging else 0
+        tic = quick_time()
         Q, R = la.qr(A_ske, overwrite_a=True, mode='economic')
-        toc = time.time() if logging else 0
-        time_factor = toc - tic
-        self.log['time_factor'] = time_factor
+        self.log['time_factor'] = quick_time() - tic
 
         # Sketch-and-solve type preprocessing
-        tic = time.time() if logging else 0
+        tic = quick_time()
         b_ske = S @ b
         z_ske = Q.T @ b_ske
         x_ske = la.solve_triangular(R, z_ske, lower=False)
         if np.linalg.norm(A @ x_ske - b) >= np.linalg.norm(b):
             z_ske = None
-        toc = time.time() if logging else 0
-        time_presolve = toc - tic
-        self.log['time_presolve'] = time_presolve
+        self.log['time_presolve'] = quick_time() - tic
 
         # Iterative phase
-        tic = time.time() if logging else 0
+        tic = quick_time()
         res = self.iterative_solver(A, b, None, 0.0, tol, iter_lim, R, True, z_ske)
-        toc = time.time() if logging else 0
-        time_iterate = toc - tic
-        self.log['time_iterate'] = time_iterate
+        self.log['time_iterate'] = quick_time() - tic
 
-        iters = res[3]
-        # Record a vector of cumulative times to (1) sketch and factor, and
-        # (2) take an individual step in LSQR (amortized!).
-        #
-        # Amortizing the time taken by a single step of LSQR is reasonable,
-        # because convergence behavior can be seen by how the normal
-        # equation error decays from one iteration to the next.
-        time_setup = time_sketch + time_factor
-        amortized = np.linspace(0, time_iterate, iters, endpoint=True)
-        cumulative = time_setup + time_presolve + amortized
-        times = np.concatenate(([time_setup], cumulative))
-        self.log['times'] = times
+        if logging:
+            iters = res[3]
+            # Record a vector of cumulative times to (1) sketch and factor, and
+            # (2) take an individual step in LSQR (amortized!).
+            #
+            # Amortizing the time taken by a single step of LSQR is reasonable,
+            # because convergence behavior can be seen by how the normal
+            # equation error decays from one iteration to the next.
+            time_setup = self.log['time_sketch']
+            time_setup += self.log['time_factor']
+            amortized = np.linspace(0, self.log['time_iterate'],
+                                    iters, endpoint=True)
+            cumulative = time_setup + self.log['time_presolve'] + amortized
+            times = np.concatenate(([time_setup], cumulative))
+            self.log['times'] = times
 
-        arnorms = res[8][:iters]
-        # Record a vector of (preconditioned) normal equation errors. Treat
-        # the zero vector as a theoretically valid initialization point which
-        # we would use before the "solve" phase of "sketch-and-solve".
-        ar0 = A.T @ b
-        ar0 = la.solve_triangular(R, ar0, 'T', lower=False, overwrite_b=True)
-        ar0norm = la.norm(ar0)
-        arnorms = np.concatenate(([ar0norm], arnorms))
-        self.log['x'] = res[0]
-        self.log['arnorms'] = arnorms
+            arnorms = res[8][:iters]
+            # Record a vector of (preconditioned) normal equation errors. Treat
+            # the zero vector as a theoretically valid initialization point which
+            # we would use before the "solve" phase of "sketch-and-solve".
+            ar0 = A.T @ b
+            ar0 = la.solve_triangular(R, ar0, 'T', lower=False, overwrite_b=True)
+            ar0norm = la.norm(ar0)
+            arnorms = np.concatenate(([ar0norm], arnorms))
+            self.log['x'] = res[0]
+            self.log['arnorms'] = arnorms
+
         return res[0]
 
 
@@ -308,40 +306,42 @@ class SPO1(OverLstsqSolver):
             assert tol < np.inf
         rng = np.random.default_rng(rng)
 
+        if logging:
+            quick_time = time.time()
+        else:
+            quick_time = lambda: 0.0
+
         # Sketch the data matrix
-        tic = time.time() if logging else 0
+        tic = quick_time()
         S = self.sketch_op_gen(d, n_rows, rng)
         A_ske = S @ A
-        toc = time.time() if logging else 0
-        self.log['time_sketch'] = toc - tic
+        self.log['time_sketch'] = quick_time() - tic
 
         # Factor the sketch
         #   We also measure the time to scale the right singular vectors
         #   as needed for the preconditioner. SPO3 doesn't have a
         #   directly comparable cost.
-        tic = time.time() if logging else 0
+        tic = quick_time()
         U, sigma, Vh = la.svd(A_ske, overwrite_a=True, check_finite=False,
                               full_matrices=False)
         eps = np.finfo(float).eps
         rank = np.count_nonzero(sigma > sigma[0] * n_cols * eps)
         N = Vh[:rank, :].T / sigma[:rank]
-        toc = time.time() if logging else 0
-        self.log['time_factor'] = toc - tic
+        self.log['time_factor'] = quick_time() - tic
 
         if self.smart_init:
-            tic = time.time() if logging else 0
+            tic = quick_time()
             b_ske = S @ b
             z_ske = U[:, :rank].T @ b_ske
             x_ske = N @ z_ske
             b_remainder = b - A @ x_ske
             if la.norm(b_remainder, ord=2) >= la.norm(b, ord=2):
                 z_ske = None
-            toc = time.time() if logging else 0
-            self.log['time_presolve'] = toc - tic
+            self.log['time_presolve'] = quick_time() - tic
 
-            tic = time.time() if logging else 0
+            tic = quick_time()
             res = self.iterative_solver(A, b, None, 0.0, tol, iter_lim, N, False, z_ske)
-            toc = time.time() if logging else 0
+            toc = quick_time()
             self.log['time_iterate'] = toc - tic
             x_star = res[0]
         else:
@@ -349,33 +349,36 @@ class SPO1(OverLstsqSolver):
             self.log['time_presolve'] = 0
 
             # Iterative phase
-            tic = time.time() if logging else 0
+            tic = quick_time()
             res = self.iterative_solver(A, b, None, 0.0, tol, iter_lim, N, False, None)
-            toc = time.time() if logging else 0
+            toc = quick_time()
             self.log['time_iterate'] = toc - tic
             x_star = res[0]
-        iters = res[3]
-        # Record a vector of cumulative times to (1) sketch and factor, and
-        # (2) take an individual step in LSQR (amortized!).
-        #
-        # Amortizing the time taken by a single step of LSQR is reasonable,
-        # because convergence behavior can be seen by how the normal
-        # equation error decays from one iteration to the next.
-        time_setup = self.log['time_sketch'] + self.log['time_factor']
-        amortized = np.linspace(0, self.log['time_iterate'], iters,
-                                endpoint=True)
-        cumulative = amortized + self.log['time_presolve'] + time_setup
-        times = np.concatenate(([time_setup], cumulative))
-        self.log['times'] = times
 
-        arnorms = res[8][:iters]
-        # Record a vector of (preconditioned) normal equation errors. Treat
-        # the zero vector as a theoretically valid initialization point which
-        # we would use before the "solve" phase of "sketch-and-solve".
-        ar0 = N.T @ (A.T @ b)
-        ar0norm = la.norm(ar0)
-        arnorms = np.concatenate(([ar0norm], arnorms))
-        self.log['x'] = x_star
-        self.log['arnorms'] = arnorms
+        if logging:
+            iters = res[3]
+            # Record a vector of cumulative times to (1) sketch and factor, and
+            # (2) take an individual step in LSQR (amortized!).
+            #
+            # Amortizing the time taken by a single step of LSQR is reasonable,
+            # because convergence behavior can be seen by how the normal
+            # equation error decays from one iteration to the next.
+            time_setup = self.log['time_sketch']
+            time_setup += self.log['time_factor']
+            amortized = np.linspace(0, self.log['time_iterate'],
+                                    iters, endpoint=True)
+            cumulative = time_setup + self.log['time_presolve'] + amortized
+            times = np.concatenate(([time_setup], cumulative))
+            self.log['times'] = times
+
+            arnorms = res[8][:iters]
+            # Record a vector of (preconditioned) normal equation errors. Treat
+            # the zero vector as a theoretically valid initialization point which
+            # we would use before the "solve" phase of "sketch-and-solve".
+            ar0 = N.T @ (A.T @ b)
+            ar0norm = la.norm(ar0)
+            arnorms = np.concatenate(([ar0norm], arnorms))
+            self.log['x'] = x_star
+            self.log['arnorms'] = arnorms
 
         return x_star